refactor project structure and add I2C, PCA9685 libraries with CMake configuration
