+++ .gitignore
@@ -44,3 +44,6 @@ compile_commands.json
 CTestTestfile.cmake
 _deps
 CMakeUserPresets.json
+
+
+CarController/build/*
