+++ CarController/external/I2C/CMakeLists.txt
@@ -1,0 +1,5 @@
+add_library(I2C
+    src/I2C.cpp
+)
+
+target_include_directories(I2C PUBLIC include)
