+++ CarController/external/PCA9685/CMakeLists.txt
@@ -1,0 +1,5 @@
+add_library(PCA9685
+    src/PCA9685.cpp
+)
+
+target_include_directories(PCA9685 PUBLIC include)
