+++ docs/rules/MainRule.json
